[IMP] spreadsheet: Remove dispatch from CoreViewPlugin
The CoreView plugins have a local state that is directly derived from the
core data. As such, they should not have any impact on plugins other than
themselves, especially not on core plugins.

This revision removes the dispatch fro the core view plugins altogether
as they don't and should not dispatch anything.

Task: 4241403
