+++ addons/spreadsheet/static/src/@types/getters.d.ts
@@ -2,8 +2,8 @@ import { CorePlugin, Model, UID } from "@odoo/o-spreadsheet";
 import { ChartOdooMenuPlugin, OdooChartCorePlugin, OdooChartUIPlugin } from "@spreadsheet/chart";
 import { CurrencyPlugin } from "@spreadsheet/currency/plugins/currency";
 import { AccountingPlugin } from "addons/spreadsheet_account/static/src/plugins/accounting_plugin";
-import { GlobalFiltersCorePlugin, GlobalFiltersUIPlugin } from "@spreadsheet/global_filters";
-import { ListCorePlugin, ListUIPlugin } from "@spreadsheet/list";
+import { GlobalFiltersCorePlugin, GlobalFiltersCoreViewPlugin } from "@spreadsheet/global_filters";
+import { ListCorePlugin, ListCoreViewPlugin } from "@spreadsheet/list";
 import { IrMenuPlugin } from "@spreadsheet/ir_ui_menu/ir_ui_menu_plugin";
 import { PivotOdooCorePlugin } from "@spreadsheet/pivot";
 import { PivotCoreGlobalFilterPlugin } from "@spreadsheet/pivot/plugins/pivot_core_global_filter_plugin";
@@ -66,8 +66,8 @@ declare module "@spreadsheet" {
 
     interface OdooGetters extends Getters {}
     interface OdooGetters extends OdooCoreGetters {}
-    interface OdooGetters extends PluginGetters<typeof GlobalFiltersUIPlugin> {}
-    interface OdooGetters extends PluginGetters<typeof ListUIPlugin> {}
+    interface OdooGetters extends PluginGetters<typeof GlobalFiltersCoreViewPlugin> {}
+    interface OdooGetters extends PluginGetters<typeof ListCoreViewPlugin> {}
     interface OdooGetters extends PluginGetters<typeof OdooChartUIPlugin> {}
     interface OdooGetters extends PluginGetters<typeof CurrencyPlugin> {}
     interface OdooGetters extends PluginGetters<typeof AccountingPlugin> {}
