+++ addons/spreadsheet/static/src/global_filters/index.js
@@ -2,6 +2,7 @@ import * as spreadsheet from "@odoo/o-spreadsheet";
 
 import { GlobalFiltersUIPlugin } from "./plugins/global_filters_ui_plugin";
 import { GlobalFiltersCorePlugin } from "./plugins/global_filters_core_plugin";
+import { GlobalFiltersCoreViewPlugin } from "./plugins/global_filters_core_view_plugin";
 const { inverseCommandRegistry } = spreadsheet.registries;
 
 function identity(cmd) {
@@ -54,4 +55,4 @@ inverseCommandRegistry
         ];
     });
 
-export { GlobalFiltersCorePlugin, GlobalFiltersUIPlugin };
+export { GlobalFiltersCorePlugin, GlobalFiltersCoreViewPlugin, GlobalFiltersUIPlugin };
