+++ addons/spreadsheet/static/src/index.js
@@ -15,11 +15,15 @@
 
 /** TODO: Introduce a position parameter to the plugin registry in order to load them in a specific order */
 import * as spreadsheet from "@odoo/o-spreadsheet";
-const { corePluginRegistry, coreViewsPluginRegistry } = spreadsheet.registries;
+const { corePluginRegistry, coreViewsPluginRegistry, featurePluginRegistry } = spreadsheet.registries;
 
-import { GlobalFiltersCorePlugin, GlobalFiltersUIPlugin } from "@spreadsheet/global_filters/index";
-import { PivotOdooCorePlugin, PivotUIGlobalFilterPlugin } from "@spreadsheet/pivot/index"; // list depends on filter for its getters
-import { ListCorePlugin, ListUIPlugin } from "@spreadsheet/list/index"; // pivot depends on filter for its getters
+import {
+  GlobalFiltersCorePlugin,
+  GlobalFiltersUIPlugin,
+  GlobalFiltersCoreViewPlugin,
+} from "@spreadsheet/global_filters/index";
+import { PivotOdooCorePlugin, PivotCoreViewGlobalFilterPlugin, PivotUIGlobalFilterPlugin } from "@spreadsheet/pivot/index"; // list depends on filter for its getters
+import { ListCorePlugin, ListCoreViewPlugin, ListUIPlugin } from "@spreadsheet/list/index"; // pivot depends on filter for its getters
 import {
     ChartOdooMenuPlugin,
     OdooChartCorePlugin,
@@ -35,8 +39,12 @@ corePluginRegistry.add("OdooListCorePlugin", ListCorePlugin);
 corePluginRegistry.add("odooChartCorePlugin", OdooChartCorePlugin);
 corePluginRegistry.add("chartOdooMenuPlugin", ChartOdooMenuPlugin);
 
-coreViewsPluginRegistry.add("OdooGlobalFiltersUIPlugin", GlobalFiltersUIPlugin);
-coreViewsPluginRegistry.add("OdooPivotGlobalFilterUIPlugin", PivotUIGlobalFilterPlugin);
-coreViewsPluginRegistry.add("OdooListUIPlugin", ListUIPlugin);
+coreViewsPluginRegistry.add("OdooGlobalFiltersCoreViewPlugin", GlobalFiltersCoreViewPlugin);
+coreViewsPluginRegistry.add("OdooPivotGlobalFiltersCoreViewPlugin", PivotCoreViewGlobalFilterPlugin);
+coreViewsPluginRegistry.add("OdooListCoreViewPlugin", ListCoreViewPlugin);
 coreViewsPluginRegistry.add("odooChartUIPlugin", OdooChartUIPlugin);
-coreViewsPluginRegistry.add("odooPivotUIPlugin", PivotOdooUIPlugin);
+
+featurePluginRegistry.add("OdooPivotGlobalFilterUIPlugin", PivotUIGlobalFilterPlugin);
+featurePluginRegistry.add("OdooGlobalFiltersUIPlugin", GlobalFiltersUIPlugin);
+featurePluginRegistry.add("odooPivotUIPlugin", PivotOdooUIPlugin);
+featurePluginRegistry.add("odooListUIPlugin", ListUIPlugin);
