+++ addons/spreadsheet/static/src/list/index.js
@@ -5,6 +5,7 @@ import * as spreadsheet from "@odoo/o-spreadsheet";
 import "./list_functions";
 
 import { ListCorePlugin } from "@spreadsheet/list/plugins/list_core_plugin";
+import { ListCoreViewPlugin } from "@spreadsheet/list/plugins/list_core_view_plugin";
 import { ListUIPlugin } from "@spreadsheet/list/plugins/list_ui_plugin";
 
 import { SEE_RECORD_LIST, SEE_RECORD_LIST_VISIBLE } from "./list_actions";
@@ -57,4 +58,4 @@ inverseCommandRegistry
     .add("RENAME_ODOO_LIST", identity)
     .add("REMOVE_ODOO_LIST", identity);
 
-export { ListCorePlugin, ListUIPlugin };
+export { ListCorePlugin, ListCoreViewPlugin, ListUIPlugin };
